Day 1 solutions
My day 1 solution. I started off using lists to hold the set of
frequencies that had been hit before, and it took so long it didn't even
get finished after a couple of hours. So I had a look at how other
people where doing it and it was all MapSet and that's what I went with
here. It's interesting, MapSet is really easy to use for these kinds of
cases but I really had no idea that it was a thing.